pyproject.toml
[tool.poetry]
name = "kasa-plug-exporter"
version = "0.1.0"
description = ""
authors = ["Morton Jonuschat <mjonuschat@gmail.com>"]
readme = "README.md"
packages = [{include = "kasa_plug_exporter"}]

[tool.poetry.dependencies]
python = "^3.9"
python-kasa = "^0.5.0"
prometheus-client = "^0.15.0"
pyyaml = "^6.0"
click = "^8.1.3"
pydantic = "^1.10.2"
asyncio-mqtt = "^0.14.0"
bestlog = "^1.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
